Created Species module structure. Added references in projects. Remove unnessesery old empty projects
