+++ FamilyPets.Backend/FamilyPets.Backend.sln
@@ -3,13 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FamilyForPets.Infrastructure", "src\FamilyPets.Infrastructure\FamilyForPets.Infrastructure.csproj", "{5F0F2E9B-911D-452C-8A43-E81541289CB0}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FamilyForPets.API", "src\FamilyPets.API\FamilyForPets.API.csproj", "{1E3C2BC2-B9FC-41AB-8BF7-EFCEC5C3FC47}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FamilyForPets.Domain", "src\FamilyPets.Domain\FamilyForPets.Domain.csproj", "{91E83833-39A1-4C04-919B-A4A89C3C9E0D}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FamilyForPets.UseCases", "src\FamilyPets.UseCases\FamilyForPets.UseCases.csproj", "{CE0ACB85-7494-4BAD-B43A-379BE405FCD9}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{02EA681E-C7D8-13C7-8484-4AC65E1B71E8}"
 	ProjectSection(SolutionItems) = preProject
@@ -33,28 +27,24 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FamilyForPets.Volunteers.AP
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FamilyForPets.Framework", "src\FamilyForPets.Framework\FamilyForPets.Framework.csproj", "{59556A48-4D27-410E-8D99-D40CDDDC0DEE}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FamilyForPets.Species.Domain", "src\Species\FamilyForPets.Species.Domain\FamilyForPets.Species.Domain.csproj", "{AFFF725D-4DE1-427E-AE26-8095AE64CFAE}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FamilyForPets.Species.UseCases", "src\Species\FamilyForPets.Species.UseCases\FamilyForPets.Species.UseCases.csproj", "{AC1A852D-BB3A-4A8F-871B-58421653DDF8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FamilyForPets.Species.API", "src\Species\FamilyForPets.Species.API\FamilyForPets.Species.API.csproj", "{66782901-065A-4CA6-85B3-033A8F093F2A}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FamilyForPets.Species.Infrastructure", "src\Species\FamilyForPets.Species.Infrastructure\FamilyForPets.Species.Infrastructure.csproj", "{3364A8D1-51E7-4337-9DE2-90D51B74E70C}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{5F0F2E9B-911D-452C-8A43-E81541289CB0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5F0F2E9B-911D-452C-8A43-E81541289CB0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5F0F2E9B-911D-452C-8A43-E81541289CB0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5F0F2E9B-911D-452C-8A43-E81541289CB0}.Release|Any CPU.Build.0 = Release|Any CPU
 		{1E3C2BC2-B9FC-41AB-8BF7-EFCEC5C3FC47}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{1E3C2BC2-B9FC-41AB-8BF7-EFCEC5C3FC47}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1E3C2BC2-B9FC-41AB-8BF7-EFCEC5C3FC47}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1E3C2BC2-B9FC-41AB-8BF7-EFCEC5C3FC47}.Release|Any CPU.Build.0 = Release|Any CPU
-		{91E83833-39A1-4C04-919B-A4A89C3C9E0D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{91E83833-39A1-4C04-919B-A4A89C3C9E0D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{91E83833-39A1-4C04-919B-A4A89C3C9E0D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{91E83833-39A1-4C04-919B-A4A89C3C9E0D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{CE0ACB85-7494-4BAD-B43A-379BE405FCD9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CE0ACB85-7494-4BAD-B43A-379BE405FCD9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CE0ACB85-7494-4BAD-B43A-379BE405FCD9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CE0ACB85-7494-4BAD-B43A-379BE405FCD9}.Release|Any CPU.Build.0 = Release|Any CPU
 		{81DDED9D-158B-E303-5F62-77A2896D2A5A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{81DDED9D-158B-E303-5F62-77A2896D2A5A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{81DDED9D-158B-E303-5F62-77A2896D2A5A}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -83,15 +73,28 @@ Global
 		{59556A48-4D27-410E-8D99-D40CDDDC0DEE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{59556A48-4D27-410E-8D99-D40CDDDC0DEE}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{59556A48-4D27-410E-8D99-D40CDDDC0DEE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{AFFF725D-4DE1-427E-AE26-8095AE64CFAE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AFFF725D-4DE1-427E-AE26-8095AE64CFAE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AFFF725D-4DE1-427E-AE26-8095AE64CFAE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AFFF725D-4DE1-427E-AE26-8095AE64CFAE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{AC1A852D-BB3A-4A8F-871B-58421653DDF8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AC1A852D-BB3A-4A8F-871B-58421653DDF8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AC1A852D-BB3A-4A8F-871B-58421653DDF8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AC1A852D-BB3A-4A8F-871B-58421653DDF8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{66782901-065A-4CA6-85B3-033A8F093F2A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{66782901-065A-4CA6-85B3-033A8F093F2A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{66782901-065A-4CA6-85B3-033A8F093F2A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{66782901-065A-4CA6-85B3-033A8F093F2A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3364A8D1-51E7-4337-9DE2-90D51B74E70C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3364A8D1-51E7-4337-9DE2-90D51B74E70C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3364A8D1-51E7-4337-9DE2-90D51B74E70C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3364A8D1-51E7-4337-9DE2-90D51B74E70C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{5F0F2E9B-911D-452C-8A43-E81541289CB0} = {4ECC13A5-2008-4C93-8811-367772CB9E43}
 		{1E3C2BC2-B9FC-41AB-8BF7-EFCEC5C3FC47} = {4ECC13A5-2008-4C93-8811-367772CB9E43}
-		{91E83833-39A1-4C04-919B-A4A89C3C9E0D} = {4ECC13A5-2008-4C93-8811-367772CB9E43}
-		{CE0ACB85-7494-4BAD-B43A-379BE405FCD9} = {4ECC13A5-2008-4C93-8811-367772CB9E43}
 		{81DDED9D-158B-E303-5F62-77A2896D2A5A} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
 		{C1DF96DA-64F4-4560-B6ED-889AD8389C9B} = {4ECC13A5-2008-4C93-8811-367772CB9E43}
 		{F9D39D83-9843-41BF-8929-CF36E3199DA2} = {4ECC13A5-2008-4C93-8811-367772CB9E43}
@@ -99,6 +102,10 @@ Global
 		{E946D5B3-B8DC-444C-B968-6E98631744E0} = {4ECC13A5-2008-4C93-8811-367772CB9E43}
 		{F2F4228C-5E4A-4EEA-8E5C-07EFDBAE8CB2} = {4ECC13A5-2008-4C93-8811-367772CB9E43}
 		{59556A48-4D27-410E-8D99-D40CDDDC0DEE} = {4ECC13A5-2008-4C93-8811-367772CB9E43}
+		{AFFF725D-4DE1-427E-AE26-8095AE64CFAE} = {4ECC13A5-2008-4C93-8811-367772CB9E43}
+		{AC1A852D-BB3A-4A8F-871B-58421653DDF8} = {4ECC13A5-2008-4C93-8811-367772CB9E43}
+		{66782901-065A-4CA6-85B3-033A8F093F2A} = {4ECC13A5-2008-4C93-8811-367772CB9E43}
+		{3364A8D1-51E7-4337-9DE2-90D51B74E70C} = {4ECC13A5-2008-4C93-8811-367772CB9E43}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {BF0D6A16-A1B7-4BD7-B9F0-246FB31ECC14}
