--- FamilyPets.Backend/src/FamilyPets.Infrastructure/FamilyForPets.Infrastructure.csproj
@@ -1,20 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="EFCore.NamingConventions" Version="9.0.0" />
-    <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="9.0.5" />
-    <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="9.0.4" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\FamilyPets.Shared\FamilyForPets.Shared.csproj" />
-    <ProjectReference Include="..\FamilyPets.UseCases\FamilyForPets.UseCases.csproj" />
-  </ItemGroup>
-
-</Project>
