--- FamilyPets.Backend/src/FamilyPets.UseCases/FamilyForPets.UseCases.csproj
@@ -1,21 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="FluentValidation" Version="12.0.0" />
-    <PackageReference Include="FluentValidation.DependencyInjectionExtensions" Version="12.0.0" />
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="9.0.5" />
-    <PackageReference Include="Microsoft.Extensions.Logging" Version="9.0.5" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\FamilyPets.Domain\FamilyForPets.Domain.csproj" />
-    <ProjectReference Include="..\FamilyPets.Shared\FamilyForPets.Shared.csproj" />
-  </ItemGroup>
-
-</Project>
