+++ FamilyPets.Backend/src/Species/FamilyForPets.Species.Domain/FamilyForPets.Species.Domain.csproj
@@ -1,0 +1,13 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net8.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\FamilyPets.Shared\FamilyForPets.Shared.csproj" />
+  </ItemGroup>
+
+</Project>
