+++ FamilyPets.Backend/src/Species/FamilyForPets.Species.Infrastructure/ApplicationDbContext.cs
@@ -1,0 +1,32 @@
+﻿using FamilyForPets.Volunteers.Domain.Entities;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Logging;
+
+namespace FamilyForPets.Volunteers.Infrastructure
+{
+    public class ApplicationDbContext(IConfiguration configuration) : DbContext
+    {
+        private const string DATABASE = "Database";
+
+
+        public DbSet<Species> Species => Set<Species>();
+
+        protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
+        {
+            optionsBuilder.UseNpgsql(configuration.GetConnectionString(DATABASE));
+            optionsBuilder.UseSnakeCaseNamingConvention();
+            optionsBuilder.UseLoggerFactory(CreateLoggerFactory());
+            optionsBuilder.EnableSensitiveDataLogging();
+        }
+
+        protected override void OnModelCreating(ModelBuilder modelBuilder)
+        {
+            modelBuilder.ApplyConfigurationsFromAssembly(typeof(ApplicationDbContext).Assembly);
+            base.OnModelCreating(modelBuilder);
+        }
+
+        private ILoggerFactory CreateLoggerFactory() =>
+            LoggerFactory.Create(builder => { builder.AddConsole(); });
+    }
+}
