+++ FamilyPets.Backend/src/Species/FamilyForPets.Species.Infrastructure/Configurations/BreedConfiguration.cs
@@ -8,30 +8,25 @@ using FamilyForPets.Domain.VolunteerAgregate.VolunteerValueObjects;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
-namespace FamilyForPets.Infrastructure.Configurations
+namespace FamilyForPets.Species.Infrastructure.Configurations
 {
-    public class SpeciesConfiguration : IEntityTypeConfiguration<Species>
+    public class BreedConfiguration : IEntityTypeConfiguration<Breed>
     {
-        public void Configure(EntityTypeBuilder<Species> builder)
+        public void Configure(EntityTypeBuilder<Breed> builder)
         {
-            builder.ToTable("species");
+            builder.ToTable("breeds");
 
             builder.HasKey(s => s.Id);
 
             builder.Property(s => s.Id)
                 .HasConversion(
                     id => id.Value,
-                    value => SpeciesId.Create(value));
+                    value => BreedId.Create(value));
 
             builder.Property(s => s.Name)
                 .IsRequired()
                 .HasMaxLength(Species.MAX_NAME_LENGHT);
-
-            builder.HasMany(v => v.Breeds)
-                .WithOne()
-                .HasForeignKey("species_id")
-                .OnDelete(DeleteBehavior.Cascade)
-                .IsRequired();
         }
     }
+
 }
