+++ FamilyPets.Backend/src/Species/FamilyForPets.Species.Infrastructure/FamilyForPets.Species.Infrastructure.csproj
@@ -7,7 +7,8 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\FamilyPets.Shared\FamilyForPets.Shared.csproj" />
+    <ProjectReference Include="..\..\FamilyPets.Shared\FamilyForPets.Shared.csproj" />
+    <ProjectReference Include="..\FamilyForPets.Species.UseCases\FamilyForPets.Species.UseCases.csproj" />
   </ItemGroup>
 
 </Project>
