+++ FamilyPets.Backend/src/Species/FamilyForPets.Species.UseCases/FamilyForPets.Species.UseCases.csproj
@@ -1,0 +1,14 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net8.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\FamilyPets.Shared\FamilyForPets.Shared.csproj" />
+    <ProjectReference Include="..\FamilyForPets.Species.Domain\FamilyForPets.Species.Domain.csproj" />
+  </ItemGroup>
+
+</Project>
